abstract out createEntry
